requirements/base.txt
# Bleeding edge Django
django==1.8.4

# Configuration
django-environ==0.3.0
django-secure==1.0.1

# Time zones support
pytz==2015.4

# API
djangorestframework==3.2.2
django-rest-swagger==0.3.4
drf-nested-routers==0.9.0

requests==2.7.0